+++ NEWS.md
@@ -3,6 +3,10 @@
   - Safe evaluation of rank in `wald_test`
   - adding CI Length to `summary.sim` output
   - fixing bug wrt `estimate.index` in `summary.sim`
+  - updated `plot.sim`, `forestplot`, `plot.estimate`
+  - `estimate`: `type="hc3"` variance estimates
+  - `merge`, `c.estimate`: new `drop.ic` argument for dropping influence
+    functions before merging
 
 # lava 1.9.0
   - `estimate`: estimate objects can now be transformed via functions: `log`,